pyproject.toml
[project]
name = "kfood-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "uvicorn>=0.35.0",
]
